pyproject.toml
[project]
name = "sql-talk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # ChainLit
    "chainlit==2.0.601",
    "asyncpg==0.30.0",
    "greenlet==3.1.1",
    "aiohttp==3.11.11",
    "google-cloud-storage==2.19.0",
    # Vanna AI
    "vanna[pgvector,bigquery,openai]==0.7.5",
    "kaleido==0.2.1",
    "psycopg-binary==3.2.4",
    # "psycopg2-binary==2.9.10",
    "langchain_openai==0.3.2",
    "db-dtypes==1.4.0",
]
